fix: add index
Set up an index with two sliders and two text fields with no
functionality at the moment.

The idea is that the user can should be able to enter colors into
the text fields (with 6-digit hex that represents rgb values) and
move the sliders to merge them and see the color reflected on the
background of the page.